refactor: Enhanced orderService with functions to create orders and generate unique identifiers.
