+++ package.json
@@ -43,6 +43,7 @@
     "react-icons": "^5.6.0",
     "react-leaflet": "^5.0.0",
     "tailwindcss": "^4.2.2",
+    "uuid": "^14.0.0",
     "vite-plugin-pwa": "^1.3.0",
     "zod": "^4.4.3"
   },
