+++ src/features/cart/components/CartView.tsx
@@ -1,380 +1,26 @@
 import { useEffect, useMemo, useState } from 'react';
 import {
-  AlertCircle,
   AlertTriangle,
   ArrowLeft,
   ChevronDown,
   ChevronRight,
-  CreditCard,
   Loader2,
-  MapPin,
   ShoppingBag,
-  Trash2,
-  X,
 } from 'lucide-react';
-import { type User } from 'firebase/auth';
-import {
-  doc,
-  getDoc,
-  serverTimestamp,
-  writeBatch,
-  increment,
-} from 'firebase/firestore';
+import { doc, getDoc } from 'firebase/firestore';
 import { CartItemRow } from './CartItemRow';
 import { AnimatedAmount } from './AnimatedAmount';
+import { LoginModal } from './LoginModal';
+import { LocationSelectorModal } from './LocationSelectorModal';
+import { PaymentConfirmModal } from './PaymentConfirmModal';
+import { OrderSuccessModal } from './OrderSuccessModal';
+import { RemoveItemModal } from './RemoveItemModal';
 import type { CartItemWithProduct } from '../types';
 import { useCartContext, CartProvider } from './CartContext';
 import { db } from '../../../lib/firebase';
 import { useAuthUser } from '../../../hooks/useAuthUser';
-import { subscribeToUserLocations } from '../../location/services/locationService';
 import type { Location } from '../../location/types';
-
-function generateOrderId() {
-  const timestamp = Date.now().toString(36);
-  const random = Math.random().toString(36).substring(2, 6);
-  return `order-${timestamp}-${random}`;
-}
-
-function generateOrderCode(): string {
-  const digits = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
-  for (let i = digits.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [digits[i], digits[j]] = [digits[j], digits[i]];
-  }
-  return digits.slice(0, 4).join('');
-}
-
-function generateItemId(orderId: string, idx: number) {
-  return `${orderId}-item-${idx + 1}`;
-}
-
-function LoginModal({ onClose }: { onClose: () => void }) {
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center p-4"
-      role="dialog"
-      aria-modal="true"
-      aria-labelledby="login-title"
-    >
-      <div
-        className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-        onClick={onClose}
-      />
-
-      <div className="relative z-10 w-full max-w-sm rounded-2xl border border-border-light bg-card-bg-light p-6 shadow-2xl">
-        <div className="flex items-center gap-3">
-          <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-primary/10 text-primary">
-            <svg
-              aria-hidden="true"
-              viewBox="0 0 24 24"
-              className="h-5 w-5"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            >
-              <path d="M15 3h4a2 2 0 0 1 2 2v14a2 2 0 0 1-2 2h-4" />
-              <polyline points="10 17 15 12 10 7" />
-              <line x1="15" y1="12" x2="3" y2="12" />
-            </svg>
-          </div>
-          <h2 id="login-title" className="text-lg font-bold text-text-light">
-            Iniciar sesión
-          </h2>
-        </div>
-
-        <p className="mt-4 text-sm leading-relaxed text-text-light opacity-70">
-          Debes iniciar sesión para confirmar tu pedido.
-        </p>
-
-        <div className="mt-6">
-          <a
-            href="/login"
-            className="flex w-full items-center justify-center rounded-full bg-primary px-4 py-3 text-sm font-semibold text-white transition hover:opacity-90 active:scale-95"
-          >
-            Ir a iniciar sesión
-          </a>
-        </div>
-
-        <button
-          type="button"
-          onClick={onClose}
-          className="absolute right-4 top-4 inline-flex h-8 w-8 items-center justify-center rounded-full border border-border-light text-text-light opacity-70 transition hover:opacity-100"
-          aria-label="Cerrar modal"
-        >
-          <X size={16} />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function LocationSelectorModal({
-  user,
-  onClose,
-  onConfirm,
-}: {
-  user: User;
-  onClose: () => void;
-  onConfirm: (location: Location) => void;
-}) {
-  const [locations, setLocations] = useState<Location[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [selectedId, setSelectedId] = useState<string | null>(null);
-
-  useEffect(() => {
-    const unsub = subscribeToUserLocations(user.uid, (locs) => {
-      setLocations(locs);
-      setLoading(false);
-      const defaultLoc = locs.find((l) => l.isDefault);
-      if (defaultLoc?.id) {
-        setSelectedId(defaultLoc.id);
-      }
-    });
-    return () => unsub();
-  }, [user.uid]);
-
-  function handleConfirm() {
-    const selected = locations.find((l) => l.id === selectedId);
-    if (selected) {
-      onConfirm(selected);
-    }
-  }
-
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center p-4"
-      role="dialog"
-      aria-modal="true"
-      aria-labelledby="location-title"
-    >
-      <div
-        className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-        onClick={onClose}
-      />
-
-      <div className="relative z-10 w-full max-w-sm rounded-2xl border border-border-light bg-card-bg-light p-6 shadow-2xl">
-        <div className="flex items-center gap-3">
-          <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-primary/10 text-primary">
-            <MapPin size={18} />
-          </div>
-          <h2 id="location-title" className="text-lg font-bold text-text-light">
-            Seleccionar ubicación
-          </h2>
-        </div>
-
-        <div className="mt-4 flex max-h-60 flex-col gap-3 overflow-y-auto">
-          {loading ? (
-            <div className="flex flex-col items-center gap-3 py-12 text-text-light/40">
-              <Loader2 size={28} className="animate-spin text-primary/60" />
-              <p className="text-sm font-bold">Cargando ubicaciones...</p>
-            </div>
-          ) : locations.length === 0 ? (
-            <div className="flex flex-col items-center gap-3 py-12 text-text-light/40">
-              <MapPin size={32} className="text-primary/40" />
-              <p className="text-sm font-bold">No hay ubicaciones guardadas</p>
-              <a
-                href="/location"
-                className="text-xs font-semibold text-primary hover:underline"
-              >
-                Agregar ubicación
-              </a>
-            </div>
-          ) : (
-            locations.map((loc) => (
-              <button
-                key={loc.id}
-                type="button"
-                onClick={() => setSelectedId(loc.id ?? null)}
-                className={`w-full rounded-xl border p-3 text-left transition ${selectedId === loc.id
-                    ? 'border-primary bg-primary/5'
-                    : 'border-border-light hover:border-primary/50'
-                  }`}
-              >
-                <p className="text-sm font-semibold text-text-light">
-                  {loc.label}
-                </p>
-                <p className="mt-1 text-xs text-text-light opacity-60 capitalize">
-                  {loc.type}
-                  {loc.isDefault && ' (Predeterminada)'}
-                </p>
-              </button>
-            ))
-          )}
-        </div>
-
-        <div className="mt-6">
-          <button
-            type="button"
-            onClick={handleConfirm}
-            disabled={!selectedId}
-            className="flex w-full items-center justify-center rounded-full bg-primary px-4 py-3 text-sm font-semibold text-white transition hover:opacity-90 active:scale-95 disabled:cursor-not-allowed disabled:opacity-50"
-          >
-            Confirmar ubicación
-          </button>
-        </div>
-
-        <button
-          type="button"
-          onClick={onClose}
-          className="absolute right-4 top-4 inline-flex h-8 w-8 items-center justify-center rounded-full border border-border-light text-text-light opacity-70 transition hover:opacity-100"
-          aria-label="Cerrar modal"
-        >
-          <X size={16} />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function PaymentConfirmModal({
-  location,
-  total,
-  onClose,
-  onConfirm,
-  loading,
-}: {
-  location: Location;
-  total: number;
-  onClose: () => void;
-  onConfirm: () => void;
-  loading: boolean;
-}) {
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center p-4"
-      role="dialog"
-      aria-modal="true"
-      aria-labelledby="payment-title"
-    >
-      <div
-        className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-        onClick={onClose}
-      />
-
-      <div className="relative z-10 w-full max-w-sm rounded-2xl border border-border-light bg-card-bg-light p-6 shadow-2xl">
-        <div className="flex items-center gap-3">
-          <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-primary/10 text-primary">
-            <CreditCard size={18} />
-          </div>
-          <h2 id="payment-title" className="text-lg font-bold text-text-light">
-            Pago contra entrega
-          </h2>
-        </div>
-
-        <div className="mt-4 flex flex-col gap-3">
-          <div className="w-full rounded-xl border border-primary bg-primary/5 p-3 text-left">
-            <p className="text-sm font-semibold text-text-light">
-              {location.label}
-            </p>
-            <p className="mt-1 text-xs text-text-light opacity-60 capitalize">
-              {location.type}
-              {location.isDefault && ' (Predeterminada)'}
-            </p>
-          </div>
-
-          <div className="w-full rounded-xl border border-border-light p-3 text-left">
-            <p className="text-xs text-text-light opacity-60">Total a pagar</p>
-            <p className="mt-1 text-2xl font-bold text-primary">
-              Bs {total.toFixed(2)}
-            </p>
-          </div>
-
-          <div className="w-full rounded-xl border border-primary bg-primary/5 p-3 text-left">
-            <div className="flex items-start gap-3">
-              <AlertCircle className="mt-0.5 text-primary" size={20} />
-
-              <div>
-                <p className="text-sm font-semibold text-primary">Importante</p>
-
-                <p className="text-sm text-text-light opacity-80">
-                  Su pago se realizará contra entrega.
-                </p>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="mt-6">
-          <button
-            type="button"
-            onClick={onConfirm}
-            disabled={loading}
-            className="flex w-full items-center justify-center rounded-full bg-primary px-4 py-3 text-sm font-semibold text-white transition hover:opacity-90 active:scale-95 disabled:cursor-not-allowed disabled:opacity-50"
-          >
-            {loading ? (
-              <Loader2 size={16} className="animate-spin" />
-            ) : (
-              'Confirmar'
-            )}
-          </button>
-        </div>
-
-        <button
-          type="button"
-          onClick={onClose}
-          className="absolute right-4 top-4 inline-flex h-8 w-8 items-center justify-center rounded-full border border-border-light text-text-light opacity-70 transition hover:opacity-100"
-          aria-label="Cerrar modal"
-        >
-          <X size={16} />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function OrderSuccessModal({ onClose }: { onClose: () => void }) {
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center p-4"
-      role="dialog"
-      aria-modal="true"
-      aria-labelledby="success-title"
-    >
-      <div
-        className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-        onClick={onClose}
-      />
-
-      <div className="relative z-10 w-full max-w-sm rounded-2xl border border-border-light bg-card-bg-light p-6 shadow-2xl">
-        <div className="flex flex-col items-center gap-3 text-center">
-          <div className="flex h-14 w-14 shrink-0 items-center justify-center rounded-full bg-primary/10 text-primary">
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="28"
-              height="28"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            >
-              <polyline points="20 6 9 17 4 12" />
-            </svg>
-          </div>
-          <h2 id="success-title" className="text-lg font-bold text-text-light">
-            ¡Pedido confirmado!
-          </h2>
-          <p className="text-sm text-text-light opacity-70">
-            Tu pedido ha sido creado exitosamente.
-          </p>
-        </div>
-
-        <div className="mt-6">
-          <button
-            type="button"
-            onClick={onClose}
-            className="flex w-full items-center justify-center rounded-full bg-primary px-4 py-3 text-sm font-semibold text-white transition hover:opacity-90 active:scale-95"
-          >
-            Aceptar
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
+import { createOrder } from '../services/orderService';
 
 function CartViewInner() {
   const {
@@ -508,7 +154,7 @@ function CartViewInner() {
     }
   }
 
-  async function handleRemove(productId: string) {
+  function handleRemove(productId: string) {
     const item =
       itemsWithProducts.find((current) => current.productId === productId) ??
       null;
@@ -556,14 +202,12 @@ function CartViewInner() {
     setShowLocationModal(true);
   }
 
-  // Cuando el usuario confirma la ubicación, guardamos y abrimos el modal de pago
   function handleLocationSelected(location: Location) {
     setSelectedLocation(location);
     setShowLocationModal(false);
     setShowPaymentConfirmModal(true);
   }
 
-  // La creación de la orden ocurre al confirmar en el modal de pago
   async function handlePaymentConfirm() {
     if (!user || creatingOrder || !selectedLocation) return;
 
@@ -571,76 +215,12 @@ function CartViewInner() {
     setShowPaymentConfirmModal(false);
 
     try {
-      const orderId = generateOrderId();
-      const orderCode = generateOrderCode();
-      const paymentCode = orderId.replace('order', 'payment');
-      const orderItems = includedItems.map((item, idx) => {
-        const unitPrice = item.unitPrice;
-        const lineSubtotal = Number((unitPrice * item.quantity).toFixed(2));
-        return {
-          itemId: generateItemId(orderId, idx),
-          productId: item.productId,
-          productName: item.product?.name ?? item.productId,
-          unitPrice,
-          quantity: item.quantity,
-          subtotal: lineSubtotal,
-        };
-      });
-
-      const batch = writeBatch(db);
-
-      const ordersRef = doc(db, 'orders', orderId);
-      batch.set(ordersRef, {
-        orderId: orderId,
-        code: orderCode,
-        buyerId: user.uid,
-        sellerId: null,
-        status: 'CREADO',
-        incidentReason: null,
+      await createOrder({
+        user,
+        selectedLocation,
         total,
-        locationId: selectedLocation.id,
-        paymentStatus: 'PENDIENTE',
-        deliveryStatus: null,
-        deliveryId: null,
-        paymentId: paymentCode,
-        confirmedAt: null,
-        cancelledAt: null,
-        createdAt: serverTimestamp(),
-        updatedAt: serverTimestamp(),
+        includedItems: includedItems as CartItemWithProduct[],
       });
-
-      for (const orderItem of orderItems) {
-        const orderItemRef = doc(
-          db,
-          `orders/${orderId}/orderItems`,
-          orderItem.itemId
-        );
-        batch.set(orderItemRef, orderItem);
-      }
-
-      for (const orderItem of orderItems) {
-        const inventoryRef = doc(db, 'inventory', orderItem.productId);
-        batch.update(inventoryRef, {
-          stockReserved: increment(orderItem.quantity),
-          updatedAt: serverTimestamp(),
-        });
-      }
-
-      const paymentsRef = doc(db, 'payments', paymentCode);
-      batch.set(paymentsRef, {
-        paymentId: paymentCode,
-        orderId: orderId,
-        amount: total,
-        method: 'cash_on_delivery',
-        status: 'PENDIENTE',
-        registeredBy: user.uid,
-        verifiedBy: null,
-        registeredAt: serverTimestamp(),
-        verifiedAt: null,
-        updatedAt: serverTimestamp(),
-      });
-
-      await batch.commit();
 
       for (const item of items) {
         await removeItem(item.productId);
@@ -688,74 +268,6 @@ function CartViewInner() {
         >
           Comprar ahora
         </a>
-      </div>
-    );
-  }
-
-  function RemoveItemModal() {
-    if (!itemToRemove) return null;
-
-    const productName = itemToRemove.product?.name ?? itemToRemove.productId;
-
-    return (
-      <div
-        className="fixed inset-0 z-50 flex items-center justify-center p-4"
-        role="dialog"
-        aria-modal="true"
-        aria-labelledby="remove-item-title"
-      >
-        <div
-          className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-          onClick={() => setItemToRemove(null)}
-        />
-
-        <div className="relative z-10 w-full max-w-sm rounded-2xl border border-border-light bg-card-bg-light p-6 shadow-2xl">
-          <div className="flex items-center gap-3">
-            <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-red-500/15 text-red-500 dark:text-red-400">
-              <Trash2 size={18} />
-            </div>
-            <h2
-              id="remove-item-title"
-              className="text-lg font-bold text-text-light"
-            >
-              Eliminar producto
-            </h2>
-          </div>
-
-          <p className="mt-4 text-sm leading-relaxed text-text-light opacity-70">
-            ¿Está seguro/a de que quieres eliminar este producto?
-            <span className="mt-1 block font-semibold text-text-light opacity-100">
-              {productName}
-            </span>
-          </p>
-
-          <div className="mt-6 flex gap-3">
-            <button
-              type="button"
-              onClick={() => setItemToRemove(null)}
-              className="flex-1 rounded-full border border-border-light bg-transparent px-4 py-3 text-sm font-semibold text-text-light transition hover:bg-secondary-bg-light active:scale-95"
-            >
-              Cancelar
-            </button>
-
-            <button
-              type="button"
-              onClick={confirmRemoveItem}
-              className="flex-1 rounded-full bg-red-600 px-4 py-3 text-sm font-semibold text-white transition hover:bg-red-700 active:scale-95"
-            >
-              Eliminar
-            </button>
-          </div>
-
-          <button
-            type="button"
-            onClick={() => setItemToRemove(null)}
-            className="absolute right-4 top-4 inline-flex h-8 w-8 items-center justify-center rounded-full border border-border-light text-text-light opacity-70 transition hover:opacity-100"
-            aria-label="Cerrar modal"
-          >
-            <X size={16} />
-          </button>
-        </div>
       </div>
     );
   }
@@ -985,7 +497,13 @@ function CartViewInner() {
           </details>
         </aside>
 
-        <RemoveItemModal />
+        {itemToRemove && (
+          <RemoveItemModal
+            item={itemToRemove}
+            onClose={() => setItemToRemove(null)}
+            onConfirm={confirmRemoveItem}
+          />
+        )}
         {showLoginModal && (
           <LoginModal onClose={() => setShowLoginModal(false)} />
         )}
